added  logmodule to the project
